refactored server
